+++ org.jacoco.cli.test/src/org/jacoco/cli/internal/commands/ReportTest.java
@@ -68,7 +68,7 @@ public class ReportTest extends CommandTestBase {
 		ExecutionDataWriter writer = new ExecutionDataWriter(execout);
 		// Add probably invalid id for this test class:
 		writer.visitClassExecution(new ExecutionData(0x123,
-				getClass().getName().replace('.', '/'), new int[] { 1 }));
+				getClass().getName().replace('.', '/'), new long[] { 1 }));
 		execout.close();
 
 		execute("report", exec.getAbsolutePath(), "--classfiles",
