+++ org.jacoco.core.test/src/org/jacoco/core/internal/analysis/InstructionsBuilderTest.java
@@ -34,7 +34,7 @@ public class InstructionsBuilderTest {
 
 	@Before
 	public void setup() {
-		builder = new InstructionsBuilder(new int[] { 0, 1 });
+		builder = new InstructionsBuilder(new long[] { 0, 1 });
 	}
 
 	@Test
