+++ org.jacoco.core/src/org/jacoco/core/internal/analysis/SourceNodeImpl.java
@@ -114,7 +114,7 @@ public class SourceNodeImpl extends CoverageNodeImpl implements ISourceNode {
 	 * @param line
 	 *            optional line number or {@link ISourceNode#UNKNOWN_LINE}
 	 */
-	public void increment(final ICounter instructions, final int executions,
+	public void increment(final ICounter instructions, final long executions,
 			final ICounter branches, final int line) {
 		if (line != UNKNOWN_LINE) {
 			incrementLine(instructions, executions, branches, line);
@@ -125,7 +125,7 @@ public class SourceNodeImpl extends CoverageNodeImpl implements ISourceNode {
 	}
 
 	private void incrementLine(final ICounter instructions,
-			final int executions, final ICounter branches, final int line) {
+			final long executions, final ICounter branches, final int line) {
 		ensureCapacity(line, line);
 		final LineImpl l = getLine(line);
 		final int oldTotal = l.getInstructionCounter().getTotalCount();
